requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
pyyaml==6.0.1
fake-useragent==1.5.1
httpx==0.27.0
tenacity==8.3.0
loguru==0.7.2
aiofiles==23.2.1
